refactor resources
