+++ components/function-controller/internal/webhook/resources/resources.go
@@ -2,193 +2,119 @@ package resources
 
 import (
 	"context"
-	"reflect"
+	"io/ioutil"
+	"path"
 
 	"github.com/pkg/errors"
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
-	apiErrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/pointer"
-	ctlrclient "sigs.k8s.io/controller-runtime/pkg/client"
+	ctrl "sigs.k8s.io/controller-runtime"
+	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	"sigs.k8s.io/controller-runtime/pkg/source"
 )
 
-type WebhookConfig struct {
-	Prefix           string
-	Type             WebHookType
-	CABundel         []byte
-	ServiceName      string
-	ServiceNamespace string
-	Port             int32
-}
-type WebHookType string
-
-const (
-	MutatingWebhook   WebHookType = "Mutating"
-	ValidatingWebHook WebHookType = "Validating"
-
-	serverlessAPIGroup   = "serverless.kyma-project.io"
-	serverlessAPIVersion = "v1alpha1"
-
-	DefaultingWebhookName = "defaulting.webhook.serverless.kyma-project.io"
-	ValidationWebhookName = "validation.webhook.serverless.kyma-project.io"
-)
-
-func EnsureWebhookConfigurationFor(ctx context.Context, client ctlrclient.Client, config WebhookConfig, wt WebHookType) error {
-	if wt == MutatingWebhook {
-		mwhc := &admissionregistrationv1.MutatingWebhookConfiguration{}
-		if err := client.Get(ctx, types.NamespacedName{Name: DefaultingWebhookName}, mwhc); err != nil {
-			if apiErrors.IsNotFound(err) {
-				return client.Create(ctx, createMutatingWebhookConfiguration(config))
-			}
-			return errors.Wrapf(err, "failed to get defaulting MutatingWebhookConfiguration: %s", DefaultingWebhookName)
-		}
-		freshMwhc := createMutatingWebhookConfiguration(config)
-		if !reflect.DeepEqual(freshMwhc.Webhooks, mwhc.Webhooks) {
-			freshMwhc.SetResourceVersion(mwhc.GetResourceVersion())
-			return client.Update(ctx, freshMwhc)
-		}
-		return nil
+func SetupResourcesController(ctx context.Context, mgr ctrl.Manager, serviceName, serviceNamespace, secretName string) error {
+	certPath := path.Join(DefaultCertDir, CertFile)
+	certBytes, err := ioutil.ReadFile(certPath)
+	if err != nil {
+		return errors.Wrapf(err, "failed to read caBundel file: %s", certBytes)
 	}
 
-	vwhc := &admissionregistrationv1.ValidatingWebhookConfiguration{}
-	if err := client.Get(ctx, types.NamespacedName{Name: ValidationWebhookName}, vwhc); err != nil {
-		if apiErrors.IsNotFound(err) {
-			return client.Create(ctx, createValidatingWebhookConfiguration(config))
-		}
-		return errors.Wrapf(err, "failed to get validation ValidatingWebhookConfiguration: %s", ValidationWebhookName)
+	webhookConfig := WebhookConfig{
+		Type:             MutatingWebhook,
+		CABundel:         certBytes,
+		ServiceName:      serviceName,
+		ServiceNamespace: serviceNamespace,
 	}
-	freshVwhc := createValidatingWebhookConfiguration(config)
-	if !reflect.DeepEqual(freshVwhc.Webhooks, vwhc.Webhooks) {
-		freshVwhc.SetResourceVersion(vwhc.GetResourceVersion())
-		return client.Update(ctx, freshVwhc)
+	// initial webhook configuration
+
+	// We are going to talk to the API server _before_ we start the manager.
+	// Since the default manager client reads from cache, we will get an error.
+	// So, we create a "serverClient" that would read from the API directly.
+	// We only use it here, this only runs at start up, so it shouldn't bee to much for the API
+	serverClient, err := ctrlclient.New(ctrl.GetConfigOrDie(), ctrlclient.Options{})
+	if err != nil {
+		return errors.Wrap(err, "failed to create a server client")
+	}
+	if err := EnsureWebhookConfigurationFor(ctx, serverClient, webhookConfig, MutatingWebhook); err != nil {
+		return errors.Wrap(err, "failed to ensure defaulting webhook configuration")
+	}
+	if err := EnsureWebhookConfigurationFor(ctx, serverClient, webhookConfig, ValidatingWebHook); err != nil {
+		return errors.Wrap(err, "failed to ensure validating webhook configuration")
+	}
+	// watch over the configuration
+	c, err := controller.New("webhook-config-controller", mgr, controller.Options{
+		Reconciler: &resourceReconciler{
+			webhookConfig: webhookConfig,
+			client:        mgr.GetClient(),
+			secretName:    secretName,
+		},
+	})
+	if err != nil {
+		return errors.Wrap(err, "failed to create webhook-config-controller")
+	}
+
+	if err := c.Watch(&source.Kind{
+		Type: &admissionregistrationv1.ValidatingWebhookConfiguration{}},
+		&handler.EnqueueRequestForObject{},
+	); err != nil {
+		return errors.Wrap(err, "failed to watch ValidatingWebhookConfiguration")
+	}
+
+	if err := c.Watch(&source.Kind{
+		Type: &admissionregistrationv1.MutatingWebhookConfiguration{}},
+		&handler.EnqueueRequestForObject{},
+	); err != nil {
+		return errors.Wrap(err, "failed to watch MutatingWebhookConfiguration")
 	}
 	return nil
 }
 
-func createMutatingWebhookConfiguration(config WebhookConfig) *admissionregistrationv1.MutatingWebhookConfiguration {
-	failurePolicy := admissionregistrationv1.Fail
-	matchPolicy := admissionregistrationv1.Exact
-	reinvocationPolicy := admissionregistrationv1.NeverReinvocationPolicy
-	scope := admissionregistrationv1.AllScopes
-	sideEffects := admissionregistrationv1.SideEffectClassNone
-	name := "defaulting.webhook.serverless.kyma-project.io"
-
-	return &admissionregistrationv1.MutatingWebhookConfiguration{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Webhooks: []admissionregistrationv1.MutatingWebhook{
-			{
-				Name: name,
-				AdmissionReviewVersions: []string{
-					"v1beta1",
-					"v1",
-				},
-				ClientConfig: admissionregistrationv1.WebhookClientConfig{
-					CABundle: config.CABundel,
-					Service: &admissionregistrationv1.ServiceReference{
-						Namespace: config.ServiceNamespace,
-						Name:      config.ServiceName,
-						Path:      pointer.String("/defaulting"),
-						Port:      pointer.Int32(443),
-					},
-				},
-				FailurePolicy:      &failurePolicy,
-				MatchPolicy:        &matchPolicy,
-				ReinvocationPolicy: &reinvocationPolicy,
-				Rules: []admissionregistrationv1.RuleWithOperations{
-					{
-						Rule: admissionregistrationv1.Rule{
-							APIGroups: []string{
-								serverlessAPIGroup,
-							},
-							APIVersions: []string{
-								serverlessAPIVersion,
-							},
-							Resources: []string{"functions", "functions/status"},
-							Scope:     &scope,
-						},
-						Operations: []admissionregistrationv1.OperationType{
-							admissionregistrationv1.Create,
-							admissionregistrationv1.Update,
-						},
-					},
-				},
-				SideEffects:    &sideEffects,
-				TimeoutSeconds: pointer.Int32(30),
-			},
-		},
-	}
+type resourceReconciler struct {
+	webhookConfig WebhookConfig
+	secretName    string
+	client        ctrlclient.Client
 }
 
-func createValidatingWebhookConfiguration(config WebhookConfig) *admissionregistrationv1.ValidatingWebhookConfiguration {
-	failurePolicy := admissionregistrationv1.Fail
-	matchPolicy := admissionregistrationv1.Exact
-	scope := admissionregistrationv1.AllScopes
-	sideEffects := admissionregistrationv1.SideEffectClassNone
-	name := "validation.webhook.serverless.kyma-project.io"
+func (r *resourceReconciler) Reconcile(ctx context.Context, request reconcile.Request) (reconcile.Result, error) {
+	// if the request is not one of our managed resources, we bail.
+	secretNamespaced := types.NamespacedName{Name: r.secretName, Namespace: r.webhookConfig.ServiceNamespace}
+	if request.Name != DefaultingWebhookName &&
+		request.Name != ValidationWebhookName &&
+		request.NamespacedName.String() != secretNamespaced.String() {
+		return reconcile.Result{}, nil
 
-	return &admissionregistrationv1.ValidatingWebhookConfiguration{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Webhooks: []admissionregistrationv1.ValidatingWebhook{
-			{
-				Name: name,
-				AdmissionReviewVersions: []string{
-					"v1beta1",
-					"v1",
-				},
-				ClientConfig: admissionregistrationv1.WebhookClientConfig{
-					CABundle: config.CABundel,
-					Service: &admissionregistrationv1.ServiceReference{
-						Namespace: config.ServiceNamespace,
-						Name:      config.ServiceName,
-						Path:      pointer.String("/validation"),
-						Port:      pointer.Int32(443),
-					},
-				},
-				FailurePolicy: &failurePolicy,
-				MatchPolicy:   &matchPolicy,
-				Rules: []admissionregistrationv1.RuleWithOperations{
-					{
-						Rule: admissionregistrationv1.Rule{
-							APIGroups: []string{
-								serverlessAPIGroup,
-							},
-							APIVersions: []string{
-								serverlessAPIVersion,
-							},
-							Resources: []string{"functions", "functions/status"},
-							Scope:     &scope,
-						},
-						Operations: []admissionregistrationv1.OperationType{
-							admissionregistrationv1.Create,
-							admissionregistrationv1.Update,
-						},
-					},
-					{
-						Rule: admissionregistrationv1.Rule{
-							APIGroups: []string{
-								serverlessAPIGroup,
-							},
-							APIVersions: []string{
-								serverlessAPIVersion,
-							},
-							Resources: []string{"gitrepositories", "gitrepositories/status"},
-							Scope:     &scope,
-						},
-						Operations: []admissionregistrationv1.OperationType{
-							admissionregistrationv1.Create,
-							admissionregistrationv1.Update,
-							admissionregistrationv1.Delete,
-						},
-					},
-				},
-				SideEffects:    &sideEffects,
-				TimeoutSeconds: pointer.Int32(30),
-			},
-		},
 	}
+	if err := r.reconcilerWebhooks(ctx, request); err != nil {
+		return reconcile.Result{}, err
+	}
+	if err := r.reconcilerSecret(ctx, request); err != nil {
+		return reconcile.Result{}, err
+	}
+	return reconcile.Result{}, nil
+}
+
+func (r *resourceReconciler) reconcilerWebhooks(ctx context.Context, request reconcile.Request) error {
+	if request.Name == DefaultingWebhookName {
+		if err := EnsureWebhookConfigurationFor(ctx, r.client, r.webhookConfig, MutatingWebhook); err != nil {
+			return err
+		}
+	}
+	if request.Name == ValidationWebhookName {
+		if err := EnsureWebhookConfigurationFor(ctx, r.client, r.webhookConfig, ValidatingWebHook); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (r *resourceReconciler) reconcilerSecret(ctx context.Context, request reconcile.Request) error {
+	secretNamespaced := types.NamespacedName{Name: r.secretName, Namespace: r.webhookConfig.ServiceNamespace}
+	if request.NamespacedName.String() == secretNamespaced.String() {
+		return EnsureWebhookSecret(ctx, r.client, request.Name, request.Namespace, r.webhookConfig.ServiceName)
+	}
+	return nil
 }
